Cloning the vundle repo somehow causes a symlink loop.
Trying to clone the repo from the bash script fails.
Somehow this causes a symlink to be created in dotfiles/vim
which points to ~/.vim which in turn points back to dotfiles/vim.
They poop back and forth forever. I guess I just have to do that step
manually until i figure out the issue.